Remove some warnings on Visual Studio .Net

+++ IO/vtkPLY.cxx
@@ -185,7 +185,8 @@ PlyFile *vtkPLY::ply_open_for_writing(
   plyInitialize();
 
   /* tack on the extension .ply, if necessary */
-  name = (char *) myalloc (sizeof (char) * (strlen (filename) + 5));
+  name = (char *) myalloc (sizeof (char) * 
+                           (static_cast<int>(strlen (filename)) + 5));
   strcpy (name, filename);
   if (strlen (name) < 4 ||
       strcmp (name + strlen (name) - 4, ".ply") != 0)
@@ -804,7 +805,8 @@ PlyFile *vtkPLY::ply_open_for_reading(
 
   /* tack on the extension .ply, if necessary */
 
-  name = (char *) myalloc (sizeof (char) * (strlen (filename) + 5));
+  name = (char *) myalloc (sizeof (char) * 
+                           (static_cast<int>(strlen (filename) + 5)));
   strcpy (name, filename);
   if (strlen (name) < 4 ||
       strcmp (name + strlen (name) - 4, ".ply") != 0)
